# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Utility Management System API",
        "description": "API documentation for the Laravel Utility Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/contract-types": {
            "get": {
                "tags": [
                    "Contract Types"
                ],
                "summary": "Get a list of contract types",
                "description": "Returns a paginated list of all contract types.",
                "operationId": "6e708fadf1798f6e1273513f0990a4f5",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contract types retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ContractTypeResource"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 4
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contract types not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contract Types"
                ],
                "summary": "Create a new contract type",
                "description": "Creates a new contract type with specified details.",
                "operationId": "5ac1d661682fd5931d5444d144dc70e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "duration",
                                    "price",
                                    "facilities"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Standard 1-Year"
                                    },
                                    "duration": {
                                        "description": "Duration in months",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 550
                                    },
                                    "facilities": {
                                        "description": "Comma-separated list of facilities",
                                        "type": "string",
                                        "example": "Wi-Fi,Air-Con,Laundry"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contract Type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractTypeResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Creation failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/contract-types/{id}": {
            "get": {
                "tags": [
                    "Contract Types"
                ],
                "summary": "Get a single contract type",
                "description": "Returns the details of a specific contract type by its ID.",
                "operationId": "4007e3e2ec22d55048d713df0f3f62c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the contract type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractTypeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contract type not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contract Types"
                ],
                "summary": "Update an existing contract type",
                "description": "Updates the details of an existing contract type.",
                "operationId": "40004b046e8c727f91b1533f9c1ccd26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the contract type to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "duration",
                                    "price",
                                    "facilities"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Premium 2-Year"
                                    },
                                    "duration": {
                                        "description": "Duration in months",
                                        "type": "integer",
                                        "example": 24
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "facilities": {
                                        "description": "Comma-separated list of facilities",
                                        "type": "string",
                                        "example": "Wi-Fi,Air-Con,Laundry,Cleaning"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contract Type updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractTypeResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Contract Type not found"
                    },
                    "500": {
                        "description": "Update failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ContractTypeResource": {
                "title": "Contract Type Resource",
                "description": "Contract Type model representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Standard 1-Year"
                    },
                    "duration": {
                        "description": "Duration in months",
                        "type": "integer",
                        "example": 12
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "facilities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "WiFi",
                            "Air Conditioning"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}